requirements.txt
asgiref==3.3.1
Django==3.1.5
django-choices==1.7.1
djangorestframework==3.12.2
mysqlclient==2.0.1
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
